Add dataloader, unit test. Add necessary packages to requirements.txt
