+++ requirements.txt
@@ -52,6 +52,4 @@ typing_extensions==4.5.0
 tzdata==2023.3
 urllib3==1.26.15
 zipp==3.15.0
-taospy==2.7.13
-taos==1.0.1
 taos-ws-py==0.3.3
